+++ CHANGELOG.md
@@ -11,10 +11,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Testing function which sets up the App with a mocked WebSerial API is working.
 - Error message is now shown if user attempts to open already in-use port, closes #254.
+- Smart scrolling when max. scrollback size is reached, auto-scroll occurs to keep the same data in view. 
 
 ### Fixed
 
 - Test which checks written data is working again.
+- Max. scrollback buffer size value now obeyed.
 
 ## [4.1.0] - 2023-09-24